+++ lib/pages/component/gits_button/src/type_button.dart
@@ -1,1 +1,0 @@
-enum TypeButton { elevated, outlined, text }