+++ lib/pages/component/gits_text_field/src/show_status_message.dart
@@ -1,1 +1,0 @@
-enum ShowStatusMessage { none, error, info, success, warning }