+++ lib/pages/foundation/validator/validator_value_page.dart
@@ -1,5 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:gits_flutter_ui_component/pages/foundation/validator/code_snippet.dart';
+import 'package:gits_flutter_ui_component/widgets/card_highlight.dart';
 import 'package:gits_flutter_ui_component/widgets/scaffold_page.dart';
+import 'package:gits_flutter_ui_component/widgets/short_description.dart';
 
 class ValidatorValuePage extends StatelessWidget {
   const ValidatorValuePage({super.key});
@@ -9,44 +12,62 @@ class ValidatorValuePage extends StatelessWidget {
     return ScaffoldPage(
       appBar: AppBar(title: const Text('Validation')),
       children: [
-        SingleChildScrollView(
-          scrollDirection: Axis.horizontal,
-          child: DataTable(
-            columns: const [
-              DataColumn(label: Text('Name')),
-              DataColumn(label: Text('Handle')),
-            ],
-            rows: [
-              DataRow(
-                cells: [
-                  const DataCell(Text('Validator Email')),
-                  DataCell(Column(
+        CardHighlight(
+          codeSnippet: codeSnippetValidatorValue,
+          child: const ShortDescription(
+            title: 'Validator Value',
+            description: 'Create file validator value',
+          ),
+        ),
+        DataTable(
+          columns: const [
+            DataColumn(label: Text('Name')),
+            DataColumn(label: Text('Handle')),
+          ],
+          rows: [
+            DataRow(
+              cells: [
+                const DataCell(Text('Validator Email')),
+                DataCell(Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: const [
+                    Text('1. Email tidak boleh kosong'),
+                    Text('2. Email maksimal 50 karakter '),
+                    Text('3. Email tidak sah'),
+                  ],
+                )),
+              ],
+            ),
+            DataRow(
+              cells: [
+                const DataCell(Text('Validator Pin')),
+                DataCell(
+                  Column(
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: const [
-                      Text('1. Email tidak boleh kosong'),
-                      Text('2. Email maksimal 50 karakter '),
-                      Text('3. Email tidak sah'),
+                      Text('1. Pin harus diisi'),
+                      Text('2. Pin dengan format angka'),
+                      Text('3. Pin harus 6 digit'),
                     ],
-                  )),
-                ],
-              ),
-              DataRow(
-                cells: [
-                  const DataCell(Text('Validator Pin')),
-                  DataCell(
-                    Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: const [
-                        Text('1. Pin harus diisi'),
-                        Text('2. Pin dengan format angka'),
-                        Text('3. Pin harus 6 digit'),
-                      ],
-                    ),
                   ),
-                ],
-              ),
-            ],
-          ),
+                ),
+              ],
+            ),
+            DataRow(
+              cells: [
+                const DataCell(Text('Validator Confirm Pin')),
+                DataCell(
+                  Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: const [
+                      Text('1. Sudah termasuk validator pin'),
+                      Text('2. Pin tidak sesuai'),
+                    ],
+                  ),
+                ),
+              ],
+            ),
+          ],
         ),
       ],
     );